+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to **ValueStringBuilder** will be documented in this file. T
 
 ## [Unreleased]
 
+### Added
+- Added the following methods: `ReplaceRune`, `Replace`
+
 ### Changed
 - Renamed `AsRune` to `AsRuneSlice` and included some more documentation to highlight that this is a shared memory block.